Adjust Join exec abstraction
